+++ packages/cloud/src/metrics/utils/getEventsInRange.ts
@@ -1,0 +1,7 @@
+import { TrackableEvent } from '../types'
+
+export const getEventsInRange = (
+  events: TrackableEvent[],
+  fromDate: Date,
+  toDate: Date,
+) => events.filter(({ date }) => date >= fromDate && date < toDate)